backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.9.0.80
scikit-learn==1.3.0
